Write and rewrite most of it.
METHOD is a mess right now: need to put it in order.

New figures, which give a kind of outline of the results section.

Rudimentary conclusion.

Leave open the possibility of Jeff's case study.

Oh yeah, and make it a real LaTeX input file, with some of my own defines, favourite packages, etc.  Add my Makefile.