NO CHANGES
This checkin was to justify all paragraphs in the TeX file, so that further changes are easier to spot in the diff (i.e. no 500-word paragraph in which there are a couple of changes that have to be scanned for).  Best: if you edit a line, don't re-justify it.  This will make your changes very easy to spot.

...except that I also added Win to the acknowledgements.